+++ widgets/views/translationForm.php
@@ -17,7 +17,7 @@ use humhub\widgets\Button;
 <?= Html::beginTag('div', $options) ?>
 
     <?php ActiveForm::begin(['id' => 'translation-editor-form', 'action' => Url::toSave($model) , 'acknowledge' => true]) ?>
-        <div class="translation-editor-filter">
+        <div class="translation-editor-filter clearfix">
             <div class="row">
                 <div class="form-group col-md-4">
                     <label for=""><?= $model->getAttributeLabel('moduleId') ?></label>
@@ -42,12 +42,14 @@ use humhub\widgets\Button;
                 <li><?= Yii::t('TranslationModule.views_translate_index', 'Messages that no longer need translation will have their translations enclosed between a pair of "@@" marks.') ?></li>
                 <li>
                     <?= Yii::t('TranslationModule.views_translate_index', 'Message string can be used with plural forms format. Check i18n section of the documentation for details.') ?>
-                    <strong><a href="https://www.yiiframework.com/doc/guide/2.0/en/tutorial-i18n#ordinal-selection" target="_blank">(Ordinal selection)</a></strong>
+                    <strong><a href="https://www.yiiframework.com/doc/guide/2.0/en/tutorial-i18n#plural" target="_blank">(Plural pattern)</a></strong>
                 </li>
                 <li> <?= Yii::t('TranslationModule.views_translate_index', 'For more informations about translation syntax see') ?>
                     <strong><a href="http://www.yiiframework.com/doc-2.0/guide-tutorial-i18n.html" target="_blank">Yii Framework Guide I18n</a></strong>.
                 </li>
             </ul>
+
+
         </div>
 
         <div class="panel-body">
@@ -59,7 +61,11 @@ use humhub\widgets\Button;
                     'data-action-keydown' => 'search']) ?>
             </p>
 
-            <p class="clearfix" style="margin-bottom:0"><?= Button::save()->submit()->right() ?></p>
+
+
+            <p class="clearfix" style="margin-bottom:0">
+                <?= Button::save()->submit()->right() ?>
+            </p>
 
             <hr style="margin-top:0">
 
@@ -75,10 +81,12 @@ use humhub\widgets\Button;
                             <div class="pre"><?= Html::encode($original) ?></div>
                         </div>
                         <div class="form-group elem <?= $model->getTranslationFieldClass($original)?>" style="position:relative">
-                            <?= Html::textArea(TranslationLog::tid($original), $translated, ['class' => 'form-control']) ?>
+                            <?= Html::textArea(TranslationLog::tid($original), $translated, ['class' => 'form-control translation '.(empty($translated) ? 'empty' : 'translated')]) ?>
+
                             <?php if(!empty($model->getHelpBlockMessage($original))) : ?>
                                 <p class="help-block"><?= Html::encode($model->getHelpBlockMessage($original)) ?></p>
                             <?php endif; ?>
+
                             <?= Button::asLink(null, Url::toHistory($model, $original))
                                 ->icon('history')
                                 ->title(Yii::t('TranslationModule.base', 'View history'))